+++ guillotina_rediscache/files.py
@@ -39,8 +39,9 @@ class RedisFileDataManager(DBDataManager):
         redis = await self.get_redis()
         key = self.get_key()
         self._data['last_activity'] = time.time()
-        await redis.set(key, json.dumps(self._data, cls=GuillotinaJSONEncoder))
-        await redis.expire(key, self._ttl)
+        await redis.set(
+            key, json.dumps(self._data, cls=GuillotinaJSONEncoder),
+            expire=self._ttl)
 
     async def get_redis(self):
         if self._redis is None: